move regression code into ML
