+++ src/openms/include/OpenMS/ML/REGRESSION/QuadraticRegression.h
