+++ src/openms/source/ML/RANSAC/RANSACModelLinear.cpp
@@ -8,7 +8,7 @@
 
 #include <OpenMS/ML/RANSAC/RANSACModelLinear.h>
 
-#include <OpenMS/MATH/STATISTICS/LinearRegression.h>
+#include <OpenMS/ML/REGRESSION/LinearRegression.h>
 #include <numeric>
